requirements-test.txt
# for main app
osbot-fast-api-serverless
pynacl

# for pytest
pytest
pytest-cov

# for fastapi testing
httpx
requests